rename netbox clients
