+++ annet/adapters/netbox/common/status_client.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 from adaptix import Retort, name_mapping, NameStyle
 from dataclass_rest import get
 from dataclass_rest.client_protocol import FactoryProtocol
-from dataclass_rest.http.requests import RequestsClient
-from requests import Session
+
+from .client import BaseNetboxClient
 
 
 @dataclass
@@ -13,15 +13,7 @@ class Status:
     plugins: dict[str, str]
 
 
-class NetboxStatusClient(RequestsClient):
-    def __init__(self, url: str, token: str):
-        url = url.rstrip("/") + "/api/"
-        session = Session()
-        session.verify = False
-        if token:
-            session.headers["Authorization"] = f"Token {token}"
-        super().__init__(url, session)
-
+class NetboxStatusClient(BaseNetboxClient):
     def _init_response_body_factory(self) -> FactoryProtocol:
         return Retort(recipe=[
             name_mapping(name_style=NameStyle.LOWER_KEBAB)
