+++ annet/adapters/netbox/v37/api_models.py
